Major refactoring: roles and state list to pillars, pillar/id removed
roles are set up in pillars instead of grains now (more secure).

Move states to be executed per role to a dictionary in the pillar (since it is
a dictionary, it gets merged when several roles are applied).

pillar/id doesn't exist anymore, that info is in pillar/top.sls.

This allows to codify the role dependencies in pillar/role/x, as you can list
the states to be applied there, and also do an include for the pillars you
need present (eg: dotfiles' states depend on workstation's states, *and* the
pillar for workstation that sets the users).